renamed list to dukeobjects temporarily
