+++ src/main/java/dukeobjects/ExpenseList.java
@@ -1,4 +1,4 @@
-package list;
+package dukeobjects;
 
 import java.io.File;
 import java.util.List;
