+++ src/main/java/ui/Ui.java
@@ -1,7 +1,7 @@
 package ui;
 
-import list.Expense;
-import list.ExpenseList;
+import dukeobjects.Expense;
+import dukeobjects.ExpenseList;
 import exception.DukeException;
