+++ text-ui-test/EXPECTED.txt
@@ -9,23 +9,23 @@ I am Duke. What can I do for you?
 Ops, you haven't added any task!
 Got it. I've added this task:
 [T][✘] go to study
-Now you have 1 expensesList in the list.
+Now you have 1 expensesList in the dukeobjects.
 Got it. I've added this task:
 [E][✘] attend the party (at: 02/02/2019 1800 - 02/03/2019 0300)
-Now you have 2 expensesList in the list.
+Now you have 2 expensesList in the dukeobjects.
 Got it. I've added this task:
 [D][✘] return the book (by: 02/03/2019 1900)
-Now you have 3 expensesList in the list.
+Now you have 3 expensesList in the dukeobjects.
 Nice! I've marked this task as done:
 [E][✓] attend the party (at: 02/02/2019 1800 - 02/03/2019 0300)
-Here are the matching expensesList in your list:
+Here are the matching expensesList in your dukeobjects:
 1.[E][✓] attend the party (at: 02/02/2019 1800 - 02/03/2019 0300)
 No results found.
-Here are the expensesList in your list:
+Here are the expensesList in your dukeobjects:
 1.[T][✘] go to study
 2.[E][✓] attend the party (at: 02/02/2019 1800 - 02/03/2019 0300)
 3.[D][✘] return the book (by: 02/03/2019 1900)
 Noted. I've removed this task:
 [T][✘] go to study
-Now you have 2 expensesList in the list.
+Now you have 2 expensesList in the dukeobjects.
 Bye. Hope to see you again soon!
