+++ text-ui-test/input.txt
@@ -1,11 +1,11 @@
-list
+dukeobjects
 todo go to study
 event attend the party /at 02/02/2019 1800 - 02/03/2019 0300
 deadline return the book /by 02/03/2019 1900
 done 2
 find attend
 find sdsds
-list
+dukeobjects
 delete 1
 bye
